+++ cudaApplication/src/utils/load_mesh.cuh
@@ -9,7 +9,7 @@
 class LoadMesh {
 public:
     std::vector<std::variant<Triangle, Mesh>> objects;
-    LoadMesh (const Eigen::Matrix4d& transform,std::istream& input_stream);
+    LoadMesh (const Eigen::Matrix4f& transform,std::istream& input_stream);
     Mesh get_mesh() const;
 };
